Add files via upload
Two players Randomly Select one card from a Deck of cards. The card with a higher Rank  Player wins the other player's Card. if the rank is the same Card War begins. The player with a minimum amount of cards can play card war. If the minimum is not satisfied then the player loses else the Player can provide  Cards and Matching Rank With Player 2. The Player with Highest Cards will win the game.